Moved form src\main\java\ifpb\entitybase\interfaces
