+++ src/main/java/ifpb/entitybasic/interfaces/IPassword.java
@@ -1,4 +1,4 @@
-package ifpb.entitybasic;
+package ifpb.entitybasic.interfaces;
 
 public interface IPassword {
     void changePassword(String newPassword) throws Exception;
